Rename class and move it to sub-namespace
